Using a sequence to group the functionality of respawning character and camera.
(1)	A sequence node is created and connected to the Character_Respawn event.
(2)	Then 1 execution path of the sequence node is connected to the Delay node created for character respawning functionality.
(3)	A camera actor, which owns a camera component, is created to set the target view fo the camera when respawning our character.