Ragdolls and Event dispatchers
(1)	Created an event dispatcher.
(2)	Created a custom event called ‘Kill’.
(3)	Event Destroyed node and functionality removed and replaced with the Kill event.
(4)	BeginPlay event node bound to the Event to Death Dispatcher and the event dispatcher is joined with the Kill event.
(5)	Destroy call in the Event Hit functionality replaced with a call to Death Dispatcher.
(6)	In the event graph of BH_GameMode, Bind Event To On Destroyed replaced with
(7)	Enabled ragdoll on the character mesh.